Refactor phase 3a: split the audio-ingest admin API into a package
webapp/admin/audio_ingest.py was 3,180 lines and the largest Flask module
in the tree, with helpers and route handlers interleaved down its whole
length. It is now webapp/admin/audio_ingest/ — 15 modules, 14 of them
under the 400-line guidance in AGENTS.md.

The package __init__ is the compatibility shim: every name the single-file
module exposed is re-exported, and register_audio_ingest_routes(app, logger)
stays the only __all__ entry, so webapp/admin/__init__.py,
webapp/routes_settings_radio.py and app_core/websocket_push.py are
untouched.

Verification: 69 of the 73 top-level definitions are ast.dump()-identical
before and after, every non-blank line of the original lands in exactly one
module, and the full suite is green (1,997 passed, 67 xfailed). The four
deliberate differences each have a test:

- The Blueprint moved into blueprint.py, so it passes __package__ rather
  than __name__ — Flask resolves a blueprint's root path from import_name,
  and __name__ would have silently deepened it by one level.
- remove_radio_managed_audio_source reads the _audio_controller singleton.
  Importing that global by value would bind None at import time and never
  see the instance _get_audio_controller later installs, silently skipping
  the local-controller cleanup; it goes through a new
  _peek_audio_controller() accessor, mirroring _get_auto_streaming_service.
- register_audio_ingest_routes fans the caller's logger out to every
  submodule. One module had one logger global for the hook to rebind;
  fifteen modules need fifteen.
- _start_audio_sources_background drops a dead `global _streaming_lock_file`
  declaration — the function returns before it could be read, and keeping it
  would have required a circular import.

Nine monkeypatch sites across three test files were retargeted from the
package to the submodule that owns each global. Not re-exporting the mutable
globals from the shim is what made that safe: monkeypatch.setattr raises on
a missing attribute, so every stale patch failed loudly instead of becoming
a silent no-op.

tests/test_audio_ingest_package.py (16 tests) pins all of the above plus
the URL map. Its three structural guards were mutation-checked —
reintroducing the by-value import, dropping a module from the fan-out list,
and un-importing a route module each fail exactly the test that covers them.

routes_sources.py stays at 749 lines: api_get_audio_sources alone is 327,
so bringing it under the cap is a restructure rather than motion. Tracked
as phase 3a-ii in the plan, and named as a known exemption in the test so
no other module can quietly join it.

Tests: tests/test_audio_ingest_package.py, tests/test_audio_ingest.py,
tests/test_radio_audio_monitoring.py, tests/test_rbds_history_api.py,
tests/test_stream_auth.py, tests/test_release_metadata.py
